refactor(seo): drop legacy utils/seo.ts and utils/structured-data.ts
Migrates the only remaining consumer of $lib/utils/seo (stripMarkdown)
to a new $lib/seo/text.ts module, re-exported from $lib/seo. Updates
OrganizationCard.svelte and dashboard/+page.svelte import paths.
Deletes the now-unused utils/seo.ts and utils/structured-data.ts files.
svelte-check: 0 errors, 335 warnings (pre-existing baseline unchanged).
